requirements.txt
aiogram==3.19.0
Flask==3.1.0
tortoise-orm==0.24.2
